+++ src/package/themes/default/components/rating.ts
@@ -1,0 +1,16 @@
+export default {
+    gap: '0.25rem',
+    size: '1.1rem',
+    color: 'var(--vf-gutter-color)',
+    activeColor: 'var(--vf-yellow-500)',
+    hoverColor: 'var(--vf-yellow-600)',
+    focusRingShadow: 'var(--vf-states-focus-ring-shadow)',
+    focusRadius: '4px',
+    disabledOpacity: 'var(--vf-states-disabled-opacity)',
+    small: {
+        size: '0.95rem',
+    },
+    large: {
+        size: '1.35rem',
+    },
+};