+++ src/package/themes/default/index.ts
@@ -36,6 +36,7 @@ import chip from './components/chip';
 import avatar from './components/avatar';
 import slider from './components/slider';
 import stepper from './components/stepper';
+import rating from './components/rating';
 
 export default {
     ...base,
@@ -78,5 +79,6 @@ export default {
         avatar,
         slider,
         stepper,
+        rating,
     },
 };